+++ CHANGELOG.md
@@ -4,7 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [Released]
+
+## [2.0.5] - 2019-10-23
+
+### Added
+
+- Invalid file handler option
+- Drag and drop file dropzone style
+
+### Changed
+
+- File titles use ```<p>``` and ```px```
+- Custom toolbar button can now be templated and passed as a DOM node
 
 ## [2.0.4] - 2019-10-22
 
@@ -45,8 +59,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Toolbar slide animation timeout is reduced
 - Fixed vertical spacing on the toolbar icons
 - Toolbar does not auto hide when any popups are open inside
-
-## [Unreleased]
 
 ## [1.3.0] - 2019-10-01